+++ rat-text/src/text_area.rs
@@ -78,8 +78,8 @@ pub mod text_area_op;
 pub struct TextArea<'a> {
     block: Option<Block<'a>>,
     hscroll: Option<Scroll<'a>>,
-    h_max_offset: Option<usize>,
-    h_overscroll: Option<usize>,
+    h_max_offset: Option<upos_type>,
+    h_overscroll: Option<upos_type>,
     vscroll: Option<Scroll<'a>>,
 
     text_wrap: Option<TextWrap>,
@@ -342,7 +342,7 @@ impl<'a> TextArea<'a> {
     /// See [self.set_horizontal_overscroll]
     ///
     /// Default is 255.
-    pub fn set_horizontal_max_offset(mut self, offset: usize) -> Self {
+    pub fn set_horizontal_max_offset(mut self, offset: u32) -> Self {
         self.h_max_offset = Some(offset);
         self
     }
@@ -353,7 +353,7 @@ impl<'a> TextArea<'a> {
     /// See [self.set_horizontal_max_offset]
     ///
     /// Default is 16384.
-    pub fn set_horizontal_overscroll(mut self, overscroll: usize) -> Self {
+    pub fn set_horizontal_overscroll(mut self, overscroll: u32) -> Self {
         self.h_overscroll = Some(overscroll);
         self
     }
@@ -428,10 +428,10 @@ fn render_text_area(
         state.hscroll.set_overscroll_by(None);
     } else {
         if let Some(h_max_offset) = widget.h_max_offset {
-            state.hscroll.set_max_offset(h_max_offset);
+            state.hscroll.set_max_offset(h_max_offset as usize);
         }
         if let Some(h_overscroll) = widget.h_overscroll {
-            state.hscroll.set_overscroll_by(Some(h_overscroll));
+            state.hscroll.set_overscroll_by(Some(h_overscroll as usize));
         }
     }
     state.hscroll.set_page_len(state.inner.width as usize);
@@ -998,8 +998,11 @@ impl TextAreaState {
 impl TextAreaState {
     /// Current offset for scrolling.
     #[inline]
-    pub fn offset(&self) -> (usize, usize) {
-        (self.hscroll.offset(), self.vscroll.offset())
+    pub fn offset(&self) -> (upos_type, upos_type) {
+        (
+            self.hscroll.offset() as upos_type,
+            self.vscroll.offset() as upos_type,
+        )
     }
 
     /// Set the offset for scrolling.
@@ -1007,10 +1010,10 @@ impl TextAreaState {
     /// The offset uses usize, but it shouldn't exceed (u32::MAX, u32::MAX).
     /// This is due to the internal ScrollState that only knows usize.
     #[inline]
-    pub fn set_offset(&mut self, offset: (usize, usize)) -> bool {
+    pub fn set_offset(&mut self, offset: (upos_type, upos_type)) -> bool {
         self.scroll_to_cursor = false;
-        let c = self.hscroll.set_offset(offset.0);
-        let r = self.vscroll.set_offset(offset.1);
+        let c = self.hscroll.set_offset(offset.0 as usize);
+        let r = self.vscroll.set_offset(offset.1 as usize);
         r || c
     }
 
@@ -2584,23 +2587,23 @@ impl TextAreaState {
     ///
     /// This is set to `len_lines - page_size` for shift-mode,
     /// and to `len_lines` for any wrapping-mode.
-    pub fn vertical_max_offset(&self) -> usize {
-        self.vscroll.max_offset()
+    pub fn vertical_max_offset(&self) -> upos_type {
+        self.vscroll.max_offset() as upos_type
     }
 
     /// Current vertical offset.
-    pub fn vertical_offset(&self) -> usize {
-        self.vscroll.offset()
+    pub fn vertical_offset(&self) -> upos_type {
+        self.vscroll.offset() as upos_type
     }
 
     /// Rendered height of the widget.
-    pub fn vertical_page(&self) -> usize {
-        self.vscroll.page_len()
+    pub fn vertical_page(&self) -> upos_type {
+        self.vscroll.page_len() as upos_type
     }
 
     /// Suggested scroll per scroll-event.
-    pub fn vertical_scroll(&self) -> usize {
-        self.vscroll.scroll_by()
+    pub fn vertical_scroll(&self) -> upos_type {
+        self.vscroll.scroll_by() as upos_type
     }
 
     /// Maximum horizontal offset.
@@ -2611,23 +2614,23 @@ impl TextAreaState {
     /// doesn't try to find an overall text-width.
     ///
     /// It __will__ be used to render the scrollbar though.
-    pub fn horizontal_max_offset(&self) -> usize {
-        self.hscroll.max_offset()
+    pub fn horizontal_max_offset(&self) -> upos_type {
+        self.hscroll.max_offset() as upos_type
     }
 
     /// Current horizontal offset.
-    pub fn horizontal_offset(&self) -> usize {
-        self.hscroll.offset()
+    pub fn horizontal_offset(&self) -> upos_type {
+        self.hscroll.offset() as upos_type
     }
 
     /// Rendered width of the text-area.
-    pub fn horizontal_page(&self) -> usize {
-        self.hscroll.page_len()
+    pub fn horizontal_page(&self) -> upos_type {
+        self.hscroll.page_len() as upos_type
     }
 
     /// Suggested scroll-by per scroll-event.
-    pub fn horizontal_scroll(&self) -> usize {
-        self.hscroll.scroll_by()
+    pub fn horizontal_scroll(&self) -> upos_type {
+        self.hscroll.scroll_by() as upos_type
     }
 
     /// Change the vertical offset.
@@ -2636,10 +2639,10 @@ impl TextAreaState {
     /// Return
     ///
     /// `true` if the offset changed at all.
-    pub fn set_vertical_offset(&mut self, row_offset: usize) -> bool {
+    pub fn set_vertical_offset(&mut self, row_offset: upos_type) -> bool {
         self.scroll_to_cursor = false;
         self.sub_row_offset = 0;
-        self.vscroll.set_offset(row_offset)
+        self.vscroll.set_offset(row_offset as usize)
     }
 
     /// Change the horizontal offset.
@@ -2650,9 +2653,9 @@ impl TextAreaState {
     /// Return
     ///
     /// `true` if the offset changed at all.
-    pub fn set_horizontal_offset(&mut self, col_offset: usize) -> bool {
+    pub fn set_horizontal_offset(&mut self, col_offset: upos_type) -> bool {
         self.scroll_to_cursor = false;
-        self.hscroll.set_offset(col_offset)
+        self.hscroll.set_offset(col_offset as usize)
     }
 
     /// Scrolls to make the given position visible.
@@ -2694,7 +2697,7 @@ impl TextAreaState {
                         ox
                     };
 
-                    self.set_offset((nox as usize, noy as usize));
+                    self.set_offset((nox, noy));
                     self.set_sub_row_offset(0);
                 }
                 TextWrap::Hard | TextWrap::Word(_) => {
@@ -2711,13 +2714,13 @@ impl TextAreaState {
                             if pos_row < off_row && pos_row >= off_row.saturating_sub(page) {
                                 let noff_row = pos_row;
                                 let (nsub_row_offset, noy) = self.stc_sub_row_offset(scr, noff_row);
-                                self.set_offset((0, noy as usize));
+                                self.set_offset((0, noy));
                                 self.set_sub_row_offset(nsub_row_offset);
                                 break 'f;
                             } else if pos_row >= off_row + page && pos_row < off_row + 2 * page {
                                 let noff_row = pos_row.saturating_sub(page.saturating_sub(1));
                                 let (nsub_row_offset, noy) = self.stc_sub_row_offset(scr, noff_row);
-                                self.set_offset((0, noy as usize));
+                                self.set_offset((0, noy));
                                 self.set_sub_row_offset(nsub_row_offset);
                                 break 'f;
                             } else if pos_row >= off_row && pos_row < off_row + page {
@@ -2732,10 +2735,10 @@ impl TextAreaState {
                     if let Some(alt_scr_row) = self.stc_screen_row(alt_scr, pos) {
                         let noff_row = alt_scr_row.saturating_sub(page * 5 / 10);
                         let (nsub_row_offset, noy) = self.stc_sub_row_offset(alt_scr, noff_row);
-                        self.set_offset((0, noy as usize));
+                        self.set_offset((0, noy));
                         self.set_sub_row_offset(nsub_row_offset);
                     } else {
-                        self.set_offset((0, pos.y as usize));
+                        self.set_offset((0, pos.y));
                         self.set_sub_row_offset(0);
                     }
                 }
@@ -2753,14 +2756,14 @@ impl TextAreaState {
     /// Return
     ///
     /// `true` if the offset changed.
-    pub fn scroll_to_row(&mut self, pos: usize) -> bool {
+    pub fn scroll_to_row(&mut self, pos: upos_type) -> bool {
         self.scroll_to_cursor = false;
 
         match self.text_wrap {
-            TextWrap::Shift => self.vscroll.scroll_to_pos(pos),
-            TextWrap::Hard | TextWrap::Word(_) => {
-                self.vscroll.set_offset(self.vscroll.limited_offset(pos))
-            }
+            TextWrap::Shift => self.vscroll.scroll_to_pos(pos as usize),
+            TextWrap::Hard | TextWrap::Word(_) => self
+                .vscroll
+                .set_offset(self.vscroll.limited_offset(pos as usize)),
         }
     }
 
@@ -2771,9 +2774,9 @@ impl TextAreaState {
     /// Return
     ///
     /// `true` if the offset changed.
-    pub fn scroll_to_col(&mut self, pos: usize) -> bool {
+    pub fn scroll_to_col(&mut self, pos: upos_type) -> bool {
         self.scroll_to_cursor = false;
-        self.hscroll.set_offset(pos)
+        self.hscroll.set_offset(pos as usize)
     }
 
     /// Scroll up by `delta` rows.
@@ -2781,7 +2784,7 @@ impl TextAreaState {
     /// Return
     ///
     /// `true` if the offset changes.
-    pub fn scroll_up(&mut self, delta: usize) -> bool {
+    pub fn scroll_up(&mut self, delta: upos_type) -> bool {
         if let Some(pos) = self.relative_screen_to_pos((0, -(delta as i16))) {
             self.sub_row_offset = pos.x;
             self.vscroll.set_offset(pos.y as usize);
@@ -2796,7 +2799,7 @@ impl TextAreaState {
     /// Return
     ///
     /// `true` if the offset changes.
-    pub fn scroll_down(&mut self, delta: usize) -> bool {
+    pub fn scroll_down(&mut self, delta: upos_type) -> bool {
         if let Some(pos) = self.relative_screen_to_pos((0, delta as i16)) {
             self.sub_row_offset = pos.x;
             self.vscroll.set_offset(pos.y as usize);
@@ -2815,9 +2818,9 @@ impl TextAreaState {
     /// `true` if the offset changes.
     ///
     /// TODO: Does nothing if there is any text-wrapping.
-    pub fn scroll_left(&mut self, delta: usize) -> bool {
+    pub fn scroll_left(&mut self, delta: upos_type) -> bool {
         self.hscroll
-            .set_offset(self.hscroll.offset.saturating_add(delta))
+            .set_offset(self.hscroll.offset.saturating_add(delta as usize))
     }
 
     /// Scroll right by `delta` columns.
@@ -2829,9 +2832,9 @@ impl TextAreaState {
     /// `true`if the offset changes.
     ///
     /// TODO: Does nothing if there is any text-wrapping.
-    pub fn scroll_right(&mut self, delta: usize) -> bool {
+    pub fn scroll_right(&mut self, delta: upos_type) -> bool {
         self.hscroll
-            .set_offset(self.hscroll.offset.saturating_sub(delta))
+            .set_offset(self.hscroll.offset.saturating_sub(delta as usize))
     }
 
     #[deprecated(since = "1.3.0", note = "not useful as is")]
@@ -3115,12 +3118,12 @@ impl HandleEvent<crossterm::event::Event, MouseOnly, TextOutcome> for TextAreaSt
             .h_scroll(&mut self.hscroll)
             .v_scroll(&mut self.vscroll);
         let r = match sas.handle(event, MouseOnly) {
-            ScrollOutcome::Up(v) => self.scroll_up(v),
-            ScrollOutcome::Down(v) => self.scroll_down(v),
-            ScrollOutcome::Left(v) => self.scroll_left(v),
-            ScrollOutcome::Right(v) => self.scroll_right(v),
-            ScrollOutcome::VPos(v) => self.scroll_to_row(v),
-            ScrollOutcome::HPos(v) => self.scroll_to_col(v),
+            ScrollOutcome::Up(v) => self.scroll_up(v as upos_type),
+            ScrollOutcome::Down(v) => self.scroll_down(v as upos_type),
+            ScrollOutcome::Left(v) => self.scroll_left(v as upos_type),
+            ScrollOutcome::Right(v) => self.scroll_right(v as upos_type),
+            ScrollOutcome::VPos(v) => self.scroll_to_row(v as upos_type),
+            ScrollOutcome::HPos(v) => self.scroll_to_col(v as upos_type),
             _ => false,
         };
         if r {